+++ src/Extension/EagerLoadingExtension.php
@@ -10,6 +10,7 @@ use ApiPlatform\Exception\RuntimeException;
 use ApiPlatform\Metadata\Operation;
 use ApiPlatform\Metadata\Property\Factory\PropertyMetadataFactoryInterface;
 use ApiPlatform\Metadata\Property\Factory\PropertyNameCollectionFactoryInterface;
+use CCMBenchmark\Ting\ApiPlatform\Mapping\ClassMetadataInfo;
 use CCMBenchmark\Ting\ApiPlatform\State\CollectionProvider;
 use CCMBenchmark\Ting\ApiPlatform\Ting\Association\AssociationType;
 use CCMBenchmark\Ting\ApiPlatform\Ting\Association\MetadataAssociation;
@@ -46,6 +47,7 @@ final class EagerLoadingExtension implements QueryCollectionExtension, QueryItem
         private readonly ClassMetadataFactoryInterface|null $classMetadataFactory = null,
         private readonly int $maxJoins = 30,
         private readonly bool $fetchPartial = false,
+        private readonly bool $forceEager = true,
     ) {
     }
 
@@ -77,6 +79,7 @@ final class EagerLoadingExtension implements QueryCollectionExtension, QueryItem
     ): void {
         $options = [];
 
+        $forceEager = $operation?->getForceEager() ?? $this->forceEager;
         $fetchPartial = $operation?->getFetchPartial() ?? $this->fetchPartial;
 
         if (! isset($context['groups']) && ! isset($context['attributes'])) {
@@ -103,6 +106,7 @@ final class EagerLoadingExtension implements QueryCollectionExtension, QueryItem
             $hydratorRelational,
             $queryNameGenerator,
             $resourceClass,
+            $forceEager,
             $fetchPartial,
             $queryBuilder->getRootAlias(),
             $options,
@@ -120,6 +124,7 @@ final class EagerLoadingExtension implements QueryCollectionExtension, QueryItem
         HydratorRelational $hydratorRelational,
         QueryNameGenerator $queryNameGenerator,
         string $resourceClass,
+        bool $forceEager,
         bool $fetchPartial,
         string $parentAlias,
         array $options = [],
@@ -154,6 +159,19 @@ final class EagerLoadingExtension implements QueryCollectionExtension, QueryItem
                 // @phpstan-ignore-next-line indeed this can be thrown by the SerializerPropertyMetadataFactory
             } catch (ResourceClassNotFoundException) {
                 // skip associations that are not resource classes
+                continue;
+            }
+
+            if (!isset($association['fetch'])) {
+                $association['fetch'] = null;
+            }
+
+            if (
+                // Always skip extra lazy associations
+                ClassMetadataInfo::FETCH_EXTRA_LAZY === $association['fetch']
+                // We don't want to interfere with ting on this association
+                || (false === $forceEager && ClassMetadataInfo::FETCH_EAGER !== $association['fetch'])
+            ) {
                 continue;
             }
 
@@ -265,6 +283,7 @@ final class EagerLoadingExtension implements QueryCollectionExtension, QueryItem
                 $hydratorRelational,
                 $queryNameGenerator,
                 $association['targetEntity'],
+                $forceEager,
                 $fetchPartial,
                 $associationAlias,
                 $options,