+++ src/Mapping/ClassMetadataInfo.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace CCMBenchmark\Ting\ApiPlatform\Mapping;
+
+class ClassMetadataInfo
+{
+    public const FETCH_LAZY = 2;
+
+    public const FETCH_EAGER = 3;
+
+    public const FETCH_EXTRA_LAZY = 4;
+}